Moved BuiltinModel and BuiltinOrCustomModel to their own module with the cli dir
